+++ README.md
@@ -1,10 +1,18 @@
 # tensorflow-research
+
 > Work in progress!
 
 ## Goal
+
 The goal of this repository is to research TensorFlow for machine vision industrial usage. <br/>
 It will provide easy to use methods to train models from scratch and also fine tune them afterwards.
 
 ## Use
+
 The "train.py" script is consuming a configuration file (you will find some predefined in "config" folder).
+
 > python src/train.py config/sample.json
+
+## Models
+### Segmentation
+- [Custom/Vanilla/Satellite Unet](https://github.com/karolzak/keras-unet)
