+++ src/models/segmentation/small_unet_model.py
