+++ src/utils/model_creater.py
@@ -1,7 +1,8 @@
 from models.anomaly_detection.advanced_model import AdvancedModel
 from models.anomaly_detection.fast_model import FastModel
-from models.anomaly_detection.satellite_unet_model import SatelliteUnetModel
-from models.anomaly_detection.small_unet_model import SmallUnetModel
+from models.segmentation.satellite_unet_model import SatelliteUnetModel
+from models.segmentation.custom_unet_model import CustomUnetModel
+from models.segmentation.small_unet_model import SmallUnetModel
 
 
 def create_model(config):
@@ -13,6 +14,8 @@ def create_model(config):
         model_container = SmallUnetModel(config)
     elif config.model == "satellite_unet":
         model_container = SatelliteUnetModel(config)
+    elif config.model == "custom_unet":
+        model_container = CustomUnetModel(config)
     else:
         TypeError
